Major Ego Gun Rework (#3273)
* start of the changes

* reload changes and charge up stuff

stuff

* reload and charge up changes

* changes

* ranged_Weapon stuff

* ports alternate fireing modes

* Update _ranged.dm

* suff

* somw other changes

* did some waw guns + he guns

* more stuff

* changes

* basic changes

* finished up to aleph

* changes

* Update waw.dm

* teth stuff

* Update teth.dm

* gaze changes

* Update he.dm

* changes

* fixes

* Renamed yangs gear to accord and some balancing and renaming stuff

* turned beams into ego bullet types

* redid loyaty's ugl falloff, hitstuff, and effects a bit more

* new line fixes

* linter stuff

* Update _ranged.dm

* Update zayin.dm

* grammar fixes and uneeded code

* Did most of the changes cox wanted

* fixed issues with fell bullet and alt shared reloading

* Nerfed Big iron scaling, fixed magic pistol not gaining benifits, and big iron and solemn eulogy now double melee force of their respective weapons

* Update waw.dm

* gave charge up a different sound

* Update waw.dm

* Fixed mobile reload not showing up and night shade being a light gun

* better reload sounds for some weapons

* Added Ammo text on the icon and some minor sound stuff with syrinx

* Added indicators for if a shared ammo/ammo dump weapon is in alt fire and some better effects for harmony

* fellscater fixes

* Adoration now no longer has iff and redid shot gun spread

* fellbullet nerf

* A dumb change that prevents map text to show up on do_item_attack_animation

* made it less stupid

* MORE HACKING SHIT TOGETHER!!!! Oh and also patching ammo on melee weapons gaining ammo off of god mode/ dead enemies

* banquet fixes

* Passive reload rebalancing

* Removed ammo_name, projectile_name  and their plural versions, They'll be readded in the future

* Redid how passive reloading works to be a semi overheat system. Also fixed some issues with burst fire and me forgetting about ammo_on_reload on some weapons

* made harmony hit scan and made sound of the star have a lower burst delay

* Redid how ebony stem worked

* did the changes cox wanted

* Added ammo colors for charged up weapons and some for banquent and fixed the funeral guns

* do_pickup_animation for ranged, redid some code for the indicators, made onreload for passive reload weapons less jank and fixed the funeral gun again

* Fixed a spread mix up for the song weapons

* ringing

* Removed Ringing losing ammo when it was used as a microphone

* fixed patriot having random spread

+++ code/modules/projectiles/projectile.dm
@@ -181,6 +181,8 @@
 	//Does this bullet ignore bulletproof?
 	var/ignore_bulletproof = FALSE			//Thumb literally cannot do anything against these goons.
 
+	var/damage_multiplier = 1
+
 /obj/projectile/Initialize()
 	. = ..()
 	decayedRange = range
@@ -358,6 +360,8 @@
  * Also, we select_target to find what to process_hit first.
  */
 /obj/projectile/proc/Impact(atom/A)
+	if(!fired)			// NEVER hit when we aren't fired
+		return
 	if(!trajectory)
 		qdel(src)
 		return FALSE
